Updated files
This is all of the updated code I have been working on. The intermediate  models are the ones that I created during the conversion to TFlite. The model that I originally create is a .pt file (a pytorch file). The tflie is stored in a seperate folder just because its the final result. The labels.txt is just the labels and the limelight needs that - I created it manually.